hl3g: HEXEDIT of libsec-ril.so to allow connections from audioserver
This is an old blob from a system which only had mediaserver. In Nougat
audioserver has been created. So you need to search for the string:

         /system/bin/mediaserver

You need to replace 'mediaserver' with 'audioserver'. Right before this
string you will find the string 'media' which is group the process
is running as, replace that with 'audio'.

chagallltetmo, ha3g, v1a3g are the only ones that needed the media group changed

Change-Id: I1ac999f8b6f9e3f992fd10b215ed42f95a8259d8
Signed-off-by: FireLord1 <firelord.xda@gmail.com>